Change Int64 to Int

+++ src/repair_eq.jl
@@ -15,7 +15,7 @@ export check_rho_1d
 """
     check_rho_1d(
         dd::IMASDD.dd;
-        time_slice::Int64=1,
+        time_slice::Int=1,
         throw_on_fail::Bool=false,
     )::Bool
 
@@ -23,7 +23,7 @@ Checks to see if rho exists and is valid in the equilibrium 1d profiles
 """
 function check_rho_1d(
     dd::IMASDD.dd;
-    time_slice::Int64=1,
+    time_slice::Int=1,
     throw_on_fail::Bool=false,
 )::Bool
     rho = dd.equilibrium.time_slice[time_slice].profiles_1d.rho_tor_norm
